spring-boot/tools/publisher/requirements.txt
google-cloud-pubsub==2.13.7
